Migrate commandlinerunner-agent to Native Build Tools agent support
Prior to this commit, the commandlinerunner-agent sample application
was failing due to improper support for using the GraalVM native image
agent.

This commit performs a migration to the Native Build Tools (NBT)
support for running with the agent in Maven.

However, this is only a partial migration since NBT currently does not
provide a way to configure Access Filters for the agent. Consequently,
the commandlinerunner-agent sample application still fails, and gh-1247
is effectively blocked until the following NBT issue is resolved.

https://github.com/graalvm/native-build-tools/issues/162

See gh-1247