Add flyway hints
This commit contains the adjustments that were made in micronaut and
quarkus to use flyway in native binaries.

The files were copied/took over from the micronaut repository which
copied/took over the files from quarkus.

A minor adjustment was made to use the flyways logger factory instead of
slf4j.

The micronaut sources can be found here:
https://github.com/micronaut-projects/micronaut-flyway

The quarkus sources can be found here:
https://github.com/quarkusio/quarkus/tree/main/extensions/flyway

The location of flyway scripts can be customized by adding the
flyway.locations property to the buildArgs of the native build
maven/gradle task. It is recognized at build time:

nativeBuild {
  ...
  buildArgs.set(listOf(
    "-Dflyway.locations=classpath:db/migration/postgres"))
}