Harmonize build-time parsing of the bean factory
Previously, BuildTimeBeanDefinitionsRegistrar was adding configuration
class parsing manually and triggered it to build the list of bean
definitions. Only certain framework callbacks were invoked.

This commit harmonizes its processing so that it is as close as possible
to what the regular runtime context would do. As a result, additional
callbacks are invoked at build time which had a subtle impact on how
bean definition types were discovered. To alleviate with that,
BeanClassBeanDefinitionPostProcessor makes sure to resolve the bean
class early if necessary.

Closes gh-1213