Add native test support in scripts
Native tests are too long to be enabled by default,
so this commit introduces --native-tests (with the
related -t shortcut) in build.sh scripts in order
to compile and run native tests in addition to the
regular native application build.

build-samples-with-native-tests.sh allows to run a full
native build for sample applications and tests.

Closes gh-1268