Prepend generated sources to test classpath for AOT tests
Prior to this commit, the Gradle plugin appended the JAR containing
generated sources and resources to the test classpath used to execute
tests on the JVM in AOT mode. This resulted in generated code not being
able to override existing code.

For example, the generated `SpringFactoriesLoader` appeared on the
classpath after the standard `SpringFactoriesLoader` present in the
`spring-core` JAR, which meant that the standard
`SpringFactoriesLoader` support was used instead of the AOT generated
variant that uses `StaticSpringFactories`.

This commit addresses this issue by prepending the generated
`*-aot-test.jar` file to the front of the test classpath.

Closes gh-1272