Migrate component processors to native configuration processors
This commit migrates the ValidatedComponentProcessor, WebComponentProcessor
and SynthesizerComputationComponentProcessor to be native configuration
processors with tests. It *also* includes an inactive translated version
of the TransactionalEventNativeComponentProcessor but we don't *seem*
to need that functionality right now so I've left it commented out - if
a full test run succeeds we could delete it.