Rework unresolved generics processing
This commit rework the processing of bean target type with unresolved
generics so that checking for such a type does not happen at runtime
anymore.

BeanDefinitionRegistrar has been changed to only set a target type if
a ResolvableType is set. And we now write it only if the target type
has generics and these are fully resolved.

Closes gh-1292