Fix support of tests with the same name in separate package
This commit improves BootstrapWriterContext so that it takes a ClassName
rather than a raw String to fork the context. Previously, the raw string
was used as the simple name of the target class, leading to conflicts
if that name was used in more than one package.

The move to ClassName clarifies the intent and provide a unique identity
for each test class. This commit also adds a sample that demonstrates
such scenario.

Closes gh-1313