Add mechanism to preserve target type
With gh-1292, we've stopped writing a resolvable type if it has
unresolved generics. This broke Spring Data repository support as a
repository may have such a signature and the context is not able to
locate the type if the target type is not set.

This is a workaround that provide a way to opt-in for such a type even
if it has unresolved generics.

See gh-1324