migrate to processBeans in SynthesizedAnnotationNativeConfigurationProcessor
this is the final usage of doWithComponents so that code has been removed now
in this commit.

This is a scary one because there is no simple guard, we have to do analysis
and so it now runs against everything to do that analysis. Maybe we could
filter it to Component but after some initial testing although it adds a few
extra entries as proxies, it isn't adding the 100s I worried it might.

Fixes #1289