Add infrastructure to exclude a bean to be written
This commit expands BeanDefinitionSelector so that it can be implemented
by a module willing to remove a bean definition that has been processed
at build-time. Typically, infrastructure beans that are no longer
needed or, potentially, bean definitions that are not supported by the
AOT engine and need to be processed again at runtime.

Closes gh-1349