Register NoOp scope if RefreshScope is used
This commit makes sure that any use of @RefreshScope safely backs off
in the native image. Previously, such an application could not start as
AOT processing was disabling the auto-configuration of the scope, but
the framework was creating a scoped proxy for it anyway.

Registering a no-op is the least invasive option and allows regular JVM
use cases to work transparently.

Closes gh-1413