Disable devtools in AOT mode
Prior to this commit, Spring Native would fail if Spring Boot Devtools
is detected on the classpath, mentioning an incompatibility with this
library.

Because Devtools is about recompiling user classes+restarting the
application context during development, this approach is currently not
compatible with AOT since AOT sources are not re-generated during this
lifecycle.

This commit now ignores completely Devtools during AOT generation and
runtime in AOT mode.

Closes gh-1419