Generate runtime sources in a distinct directory
Directories like target/generated-sources/ with Maven and
build/generated/sources with Gradle are automatically imported
by IDEs like IntelliJ IDEA and are assuming a compile classpath.

The code generated by Spring AOT is generated with the runtime
classpath, so this commit moves the generation to a dedicated
directory not imported automatically.

Maven does not differentiate compile and runtime test classpaths
but experimentation in IDEA has shown bogus behavior in debugger
when target/generated-test-sources is kept to this commit
change it to target/generated-runtime-test-sources.

Debugging those generated files now require adding them explicitly
to the classpath but our assumption is that's not a very common
need.

Closes gh-1093