Refine logback support
This commit restores and refines the LogBack substitution to relax
the check on XML files since there is already a check at build
time in LogbackHints and @aclement found a related issue.

The documentation has been updated to recommend using embedded
logback-spring.xml and the logger sample is now mentioned.

Closes gh-1505