Clear AOT sources only before AOT generation
Prior to this commit, AOT generated sources would be cleared as soon as
the AOT plugin is applied (even if no AOT taks is invoked). This would
lead to AOT-generated sources disappearing at various unexpected time
during the development lifecycle of an application.

This commit ensures that sources are only cleared right before the AOT
generation task is invoked.

Fixes gh-1552